fix: modularise DashboardAdmin
